Move use case specific code to examples directory
