--- dockerfiles/Dockerfile.split_server.jp461
@@ -1,7 +1,0 @@
-FROM nvcr.io/nvidia/l4t-pytorch:r32.7.1-pth1.9-py3
-
-WORKDIR /app
-
-COPY src .
-
-CMD ["python3", "split_training_server.py"]
