+++ examples/nvidia_demo/lib/segnet.py
