+++ examples/split_learning/.dockerignore
