+++ examples/split_learning/Dockerfile.server
@@ -1,0 +1,10 @@
+# | SDK            | Arch tag |
+# | -------------- | -------- |
+# | JetPack 5.0    | jp50     |
+# | JetPack 4.6.1  | jp461    |
+ARG ARCH_TAG=amd64
+FROM sparse/pytorch:$ARCH_TAG
+
+COPY . .
+
+CMD ["python3", "split_training_server.py"]
