+++ examples/split_learning/datasets/mnist_fashion.py
