+++ examples/split_learning/serialization.py
