+++ src/utils.py
@@ -1,12 +1,5 @@
 import sys
 
-import torch
-
-def get_device():
-    device = "cuda" if torch.cuda.is_available() else "cpu"
-    print(f"Using {device} for processing")
-    return device
-
 def use_legacy_asyncio():
     if sys.version_info >= (3, 8, 10):
         print("Using latest asyncio implementation.")
@@ -17,7 +10,3 @@ def use_legacy_asyncio():
     else:
         print("WARNING: The used Python interpreter is older than what is officially supported. This may cause some functionalities to break")
         return True
-
-def save_model(model, filepath):
-    torch.save(model.state_dict(), filepath)
-    print("Saved PyTorch Model State to {:s}".format(filepath))
